consolidate np implementation of mc
